+++ R/utils.R
@@ -1740,3 +1740,29 @@ getSliderType <- function(min, max, value) {
   }
   type[[1]]
 }
+
+# Reads the `shiny.sharedSecret` global option, and returns a function that can
+# be used to test header values for a match.
+loadSharedSecret <- function() {
+  sharedSecret <- getOption("shiny.sharedSecret")
+  if (is.null(sharedSecret)) {
+    function(x) TRUE
+  } else {
+    # We compare the digest of the two values so that their lengths are equalized
+    sharedSecret <- digest::digest(sharedSecret, "sha256", serialize = FALSE, raw = TRUE)
+    function(x) {
+      x <- digest::digest(x, "sha256", serialize = FALSE, raw = TRUE)
+      # Constant time comparison to avoid timing attacks
+      constantTimeEquals(sharedSecret, x)
+    }
+  }
+}
+
+# Compares two raw vectors of equal length for equality, in constant time
+constantTimeEquals <- function(raw1, raw2) {
+  stopifnot(is.raw(raw1))
+  stopifnot(is.raw(raw2))
+  stopifnot(length(raw1) == length(raw2))
+
+  sum(as.integer(xor(raw1, raw2))) == 0
+}